--- src/config/util/use-label.ts
@@ -1,12 +1,0 @@
-import { Emoji } from '../git_emoji_zh'
-
-export function use_label(emoji: Emoji) {
-  console.log("使用label");
-
-  return {
-    label: `${emoji.emoji} ${emoji.description}`,
-    code: emoji.code,
-    emoji: emoji.code + ' ',
-    description: '[' + emoji.name + ']',
-  }
-}
